refactor: Group `*Data*.js` under 📁`Data`
