+++ src/Plugins/Pagination.js
@@ -5,7 +5,7 @@ import EleventyBaseError from "../Errors/EleventyBaseError.js";
 import { DeepCopy } from "../Util/Merge.js";
 import { ProxyWrap } from "../Util/ProxyWrap.js";
 // import { DeepFreeze } from "../Util/DeepFreeze.js";
-import TemplateData from "../TemplateData.js";
+import TemplateData from "../Data/TemplateData.js";
 
 const { set: lodashSet, get: lodashGet, chunk: lodashChunk } = lodash;
