+++ src/TemplateCollection.js
@@ -2,7 +2,7 @@ import multimatch from "multimatch";
 import { TemplatePath } from "@11ty/eleventy-utils";
 
 import Sortable from "./Util/Sortable.js";
-import TemplateData from "./TemplateData.js";
+import TemplateData from "./Data/TemplateData.js";
 
 class TemplateCollection extends Sortable {
 	constructor() {
