+++ test/TemplateTest.js
@@ -4,7 +4,7 @@ import pretty from "pretty";
 import TOML from "@iarna/toml";
 
 import TemplateConfig from "../src/TemplateConfig.js";
-import TemplateData from "../src/TemplateData.js";
+import TemplateData from "../src/Data/TemplateData.js";
 import FileSystemSearch from "../src/FileSystemSearch.js";
 import EleventyExtensionMap from "../src/EleventyExtensionMap.js";
 import EleventyErrorUtil from "../src/Errors/EleventyErrorUtil.js";
